Implemented unit tests and introduced bug fixes found in other unit test (#59)
* Implemented a bug fix to LoginInteractor.java.

Previosuly, an error would occur if a user without any current trip would try to log in. I changed it so that if a current trip is not detected, the output data for loggedinview would just be empty.

* Added unit tests for LoadTripDetailInteractor

Edited User.java with new setter functions and removed final keywords so that user objects in the DAO are editable.